[add] TimerExecutorクラスを追加し、timerの処理を分離
