+++ src/mb/mizinkobusters/elysium/command/Start.java
@@ -2,10 +2,9 @@ package mb.mizinkobusters.elysium.command;
 
 import mb.mizinkobusters.elysium.BlockGenerator;
 import mb.mizinkobusters.elysium.Main;
-import mb.mizinkobusters.elysium.RankingUtil;
+import mb.mizinkobusters.elysium.TimerExecutor;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
-import org.bukkit.Sound;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
@@ -15,7 +14,6 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
-import org.bukkit.scheduler.BukkitRunnable;
 
 public class Start implements CommandExecutor {
 
@@ -55,68 +53,7 @@ public class Start implements CommandExecutor {
         BlockGenerator thread = new BlockGenerator();
         thread.generate();
 
-        new BukkitRunnable() {
-            int timer = 30 * 60 + 5;
-            int countdown = 0;
-
-            @Override
-            public void run() {
-                switch (timer) {
-                    case 30 * 60 + 5:
-                        countdown = 3;
-                    case 30 * 60 + 5 - 1:
-                    case 30 * 60 + 5 - 2:
-                        Bukkit.broadcastMessage(countdown + "...");
-                        countdown--;
-                        break;
-                    case 30 * 60 + 5 - 3:
-                        for (Player players : Bukkit.getOnlinePlayers()) {
-                            players.sendTitle("スタート!", "", 5, 100, 5);
-                            players.playSound(players.getLocation(), Sound.ENTITY_GENERIC_EXPLODE, 1, 1);
-                        }
-                        break;
-                    case 20 * 60 + 5:
-                    case 10 * 60 + 5:
-                        RankingUtil.showHighestPlayer();
-                        for (Player players : Bukkit.getOnlinePlayers()) {
-                            players.sendTitle("残り" + (timer / 60) + "分", "", 5, 100, 5);
-                        }
-                        break;
-                    case 5 + 5:
-                        countdown = 5;
-                    case 5 + 4:
-                    case 5 + 3:
-                    case 5 + 2:
-                    case 5 + 1:
-                        for (Player players : Bukkit.getOnlinePlayers()) {
-                            players.sendMessage(countdown + "...");
-                            players.playSound(players.getLocation(), Sound.BLOCK_ANVIL_FALL, 1, 1);
-                        }
-                        countdown--;
-                        break;
-                    case 5:
-                        for (Player players : Bukkit.getOnlinePlayers()) {
-                            players.sendTitle("終了!!", "", 5, 100, 5);
-                            players.playSound(players.getLocation(), Sound.ENTITY_GENERIC_EXPLODE, 1, 1);
-                            players.sendMessage("-------------------");
-                            players.sendMessage("今回の優勝者は...");
-                        }
-                        break;
-                    case 0:
-                        Player highestLeveler = RankingUtil.getHighestLeveler();
-                        for (Player players : Bukkit.getOnlinePlayers()) {
-                            players.playSound(players.getLocation(), Sound.ENTITY_PLAYER_LEVELUP, 1, 1);
-                            players.sendMessage("-------------------");
-                            players.sendMessage("名前: " + highestLeveler.getName());
-                            players.sendMessage("レベル: " + highestLeveler.getLevel());
-                            players.sendMessage("-------------------");
-                        }
-                        break;
-                }
-
-                timer--;
-            }
-        }.runTaskTimer(Main.getInstance(), 0, 20);
+        new TimerExecutor().runTaskTimer(Main.getInstance(), 20, 20);
 
         return true;
     }
